+++ transaction-native/src/main/java/module-info.java
@@ -1,8 +1,8 @@
-module io.ballerina.io.transaction {
+module io.ballerina.stdlib.transaction {
     requires io.ballerina.jvm;
     requires io.ballerina.lang;
     requires org.slf4j;
 
-    exports org.ballerinalang.io.transaction;
-    exports org.ballerinalang.io.transaction.compiler;
+    exports org.ballerinalang.stdlib.transaction;
+    exports org.ballerinalang.stdlib.transaction.compiler;
 }
