+++ transaction-native/src/main/java/org/ballerinalang/stdlib/transaction/CleanUpTransactionContext.java
@@ -16,21 +16,24 @@
  * under the License.
  */
 
-package org.ballerinalang.io.transaction;
+package org.ballerinalang.stdlib.transaction;
 
 import org.ballerinalang.jvm.api.values.BString;
 import org.ballerinalang.jvm.scheduling.Scheduler;
-import org.ballerinalang.jvm.transactions.TransactionResourceManager;
+import org.ballerinalang.jvm.scheduling.Strand;
+import org.ballerinalang.jvm.transactions.TransactionLocalContext;
 
 /**
- * Extern function transaction:abortResourceManagers.
+ * Extern function transaction:cleanupTransactionContext.
  *
  * @since Swan Lake
  */
-public class AbortResourceManagers {
+public class CleanUpTransactionContext {
 
-    public static boolean abortResourceManagers(BString transactionId, BString transactionBlockId) {
-        return TransactionResourceManager.getInstance().notifyAbort(Scheduler.getStrand(), transactionId.getValue(),
-                transactionBlockId.getValue(), null);
+    public static void cleanupTransactionContext(BString transactionBlockId) {
+        Strand strand = Scheduler.getStrand();
+        TransactionLocalContext transactionLocalContext = strand.currentTrxContext;
+        transactionLocalContext.removeTransactionInfo();
+        strand.removeCurrentTrxContext();
     }
 }
