+++ transaction-native/src/main/java/org/ballerinalang/stdlib/transaction/PrepareResourceManagers.java
@@ -16,18 +16,20 @@
  * under the License.
  */
 
-package org.ballerinalang.io.transaction;
+package org.ballerinalang.stdlib.transaction;
 
-import org.ballerinalang.jvm.scheduling.Scheduler;
+import org.ballerinalang.jvm.api.values.BString;
+import org.ballerinalang.jvm.transactions.TransactionResourceManager;
 
 /**
- * Extern function transaction:getAndClearFailure.
+ * Extern function transaction:prepareResourceManagers.
  *
  * @since Swan Lake
  */
-public class GetAndClearFailure {
+public class PrepareResourceManagers {
 
-    public static boolean getAndClearFailure() {
-        return Scheduler.getStrand().currentTrxContext.getAndClearFailure() != null;
+    public static boolean prepareResourceManagers(BString transactionId, BString transactionBlockId) {
+        return TransactionResourceManager.getInstance().prepare(transactionId.getValue(),
+                transactionBlockId.getValue());
     }
 }
