+++ transaction-native/src/main/java/org/ballerinalang/stdlib/transaction/RollbackTransaction.java
@@ -16,7 +16,7 @@
  * under the License.
  */
 
-package org.ballerinalang.io.transaction;
+package org.ballerinalang.stdlib.transaction;
 
 import org.ballerinalang.jvm.api.values.BString;
 import org.ballerinalang.jvm.scheduling.Scheduler;
